Restructure API implementation into separate packages
