+++ src/main/java/totemic_commons/pokefenn/apiimpl/ApiImpl.java
@@ -1,10 +1,11 @@
-package totemic_commons.pokefenn;
+package totemic_commons.pokefenn.apiimpl;
 
 import totemic_commons.pokefenn.api.TotemicAPI;
 import totemic_commons.pokefenn.api.TotemicRegistry;
 import totemic_commons.pokefenn.api.music.MusicAPI;
 import totemic_commons.pokefenn.api.totem.TotemEffectAPI;
-import totemic_commons.pokefenn.totem.TotemEffectApiImpl;
+import totemic_commons.pokefenn.apiimpl.music.MusicApiImpl;
+import totemic_commons.pokefenn.apiimpl.totem.TotemEffectApiImpl;
 
 public final class ApiImpl implements TotemicAPI.API
 {
